# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WPMgr API
  version: 0.0.0
  description: |
    Control-plane REST API for WPMgr. This is the single source of truth for
    both the Go server types and the generated TypeScript client. Real paths
    and schemas are filled in during Phase 4 (skeleton) and Phase 5 (features).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/sites/{siteId}/destinations:
    get:
      operationId: listSiteDestinations
      summary: List configured backup destinations for a site
      description: |
        ADR-036 P1 storage adapter. Returns every destination configured on the
        site (`cp` / `local` / `s3_compat`). The encrypted S3 secret is NEVER
        returned; `has_secret` reports whether one is stored.
      tags: [destinations]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: A list of destinations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDestinationList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createSiteDestination
      summary: Add a backup destination to a site
      tags: [destinations]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteDestinationCreate"
      responses:
        "201":
          description: Destination created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDestination"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/sites/{siteId}/destinations/test:
    post:
      operationId: testSiteDestination
      summary: Verify credentials for a destination without persisting them
      description: |
        Returns 200 with `{ok, message}` regardless of success/failure so the
        UI can render the operator-readable diagnostic inline. S3-compat
        kinds run a HeadBucket + PutObject + DeleteObject probe against the
        target bucket; cp/local kinds return ok=true trivially.
      tags: [destinations]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteDestinationTest"
      responses:
        "200":
          description: Test result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDestinationTestResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/sites/{siteId}/destinations/{destinationId}:
    get:
      operationId: getSiteDestination
      summary: Read one configured destination
      tags: [destinations]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - $ref: "#/components/parameters/DestinationId"
      responses:
        "200":
          description: The destination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDestination"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateSiteDestination
      summary: Update a configured destination (omit secret_key to keep it)
      tags: [destinations]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - $ref: "#/components/parameters/DestinationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteDestinationUpdate"
      responses:
        "200":
          description: The updated destination.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteDestination"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSiteDestination
      summary: Remove a configured destination
      tags: [destinations]
      parameters:
        - $ref: "#/components/parameters/SiteId"
        - $ref: "#/components/parameters/DestinationId"
      responses:
        "204":
          description: Destination deleted.
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Destination not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/sites:
    post:
      operationId: createSite
      summary: Create a site
      description: Creates a site belonging to the tenant in the request context.
      tags: [sites]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteCreate"
      responses:
        "201":
          description: Site created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "409":
          description: Site URL already exists for this tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      operationId: listSites
      summary: List sites for the current tenant
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: tag
          in: query
          required: false
          description: Filter to sites carrying this tag.
          schema:
            type: string
            maxLength: 64
      responses:
        "200":
          description: A page of sites
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteList"
...
  /api/v1/sites/{siteId}:
    get:
      operationId: getSite
      summary: Get a site by ID
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "200":
          description: The site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteSite
      summary: Delete a site
      tags: [sites]
      parameters:
        - $ref: "#/components/parameters/SiteId"
      responses:
        "204":
          description: Site deleted
        "404":
          description: Site not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
...
    SiteId:
      name: siteId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DestinationId:
      name: destinationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    AgentSignature:
      type: apiKey
      in: header
      name: X-WPMgr-Signature
      description: |
        Ed25519 signed-request authentication for agent->control-plane calls. The
        agent sends four headers: X-WPMgr-Agent-Key (its base64 Ed25519 public
        key, identifying the enrolled site), X-WPMgr-Timestamp (Unix seconds),
        X-WPMgr-Nonce (a unique per-request token), and X-WPMgr-Signature (the
        base64 Ed25519 signature over the canonical message
        "METHOD\nPATH\nTIMESTAMP\nNONCE\nhex(sha256(body))"). The control plane
        rejects stale timestamps and replayed nonces. (apiKey type is the closest
        OpenAPI 3.1 primitive for this header-carried signature scheme.)
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: object
          additionalProperties: true
...
    Site:
      type: object
      required:
        [id, tenant_id, url, name, status, wp_version, php_version, health_status, multisite, tags, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        name:
          type: string
        status:
          type: string
          enum: [pending, active, error, disabled]
        wp_version:
          type: string
        php_version:
          type: string
        health_status:
          type: string
          enum: [unknown, healthy, unreachable]
        server_info:
          type: string
        multisite:
          type: boolean
        active_theme:
          type: string
        tags:
          type: array
          items:
            type: string
        enrolled:
          type: boolean
          description: Whether an agent has enrolled this site.
        enrolled_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
        components:
          $ref: "#/components/schemas/SiteComponents"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteComponents:
      type: object
      description: Installed plugin/theme inventory pushed by the agent.
      properties:
        plugins:
          type: array
          items:
            $ref: "#/components/schemas/SiteComponent"
        themes:
          type: array
          items:
            $ref: "#/components/schemas/SiteComponent"
        core_update:
          type: object
          nullable: true
          description: When set, WordPress core has an update available.
          required: [new_version, current_version]
          properties:
            new_version:
              type: string
            current_version:
              type: string
    SiteComponent:
      type: object
      required: [slug]
      properties:
        slug:
          type: string
        name:
          type: string
        version:
          type: string
        active:
          type: boolean
        available_update:
          type: object
          nullable: true
          description: When set, an update is available for this plugin/theme.
          required: [new_version]
          properties:
            new_version:
              type: string
            package:
              type: string
              nullable: true
            tested:
              type: string
              nullable: true
            requires_php:
              type: string
              nullable: true
...
    SiteCreate:
      type: object
      required: [url, name]
      properties:
        url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        name:
          type: string
          minLength: 1
          maxLength: 200
        status:
          type: string
          enum: [pending, active, error, disabled]
          default: pending
        wp_version:
          type: string
          maxLength: 32
        php_version:
          type: string
          maxLength: 32
    SiteList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Site"
...
    SiteDestinationKind:
      type: string
      enum: [cp, local, s3_compat]
      description: |
        ADR-036 P1 destination backend type. `cp` is the WPMgr-managed bucket,
        `local` writes to wp-content/wpmgr-backups on the agent host, and
        `s3_compat` targets a customer-owned S3-compatible bucket.
    SiteDestination:
      type: object
      required:
        [id, tenant_id, site_id, kind, label, endpoint, region, bucket,
         path_prefix, access_key_id, has_secret, force_path_style, is_default,
         created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        site_id:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/SiteDestinationKind"
        label:
          type: string
        endpoint:
          type: string
        region:
          type: string
        bucket:
          type: string
        path_prefix:
          type: string
        access_key_id:
          type: string
        has_secret:
          type: boolean
          description: |
            True iff a secret key is stored. The encrypted bytes never cross
            the wire; the UI uses this to render a "Re-enter to save changes"
            hint on the secret_key field.
        force_path_style:
          type: boolean
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SiteDestinationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SiteDestination"
    SiteDestinationCreate:
      type: object
      required: [kind, label]
      properties:
        kind:
          $ref: "#/components/schemas/SiteDestinationKind"
        label:
          type: string
          minLength: 1
          maxLength: 200
        endpoint:
          type: string
        region:
          type: string
        bucket:
          type: string
        path_prefix:
          type: string
        access_key_id:
          type: string
        secret_key:
          type: string
          description: |
            PLAINTEXT secret. The CP age-encrypts it at rest before persisting;
            it is NEVER stored in clear nor read back over the API.
        force_path_style:
          type: boolean
        is_default:
          type: boolean
    SiteDestinationUpdate:
      type: object
      properties:
        label:
          type: string
        endpoint:
          type: string
        region:
          type: string
        bucket:
          type: string
        path_prefix:
          type: string
        access_key_id:
          type: string
        secret_key:
          type: string
          description: |
            Omit to keep the existing secret; non-empty replaces it.
        force_path_style:
          type: boolean
        is_default:
          type: boolean
    SiteDestinationTest:
      type: object
      required: [kind]
      properties:
        kind:
          $ref: "#/components/schemas/SiteDestinationKind"
        endpoint:
          type: string
        region:
          type: string
        bucket:
          type: string
        path_prefix:
          type: string
        access_key_id:
          type: string
        secret_key:
          type: string
        force_path_style:
          type: boolean
    SiteDestinationTestResult:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string
...